Create Street View Image Geolocation using Neural Networks
The project aims to develop a deep neural network model, for accurately determining the 
geographical location of an image within the United States. The goal is to predict the most probable 
state in which a given image was captured. This is framed as a classification problem, where the 
model will be trained using a dataset of Google Street View images. The project aims to demonstrate 
that deep neural networks can effectively geolocate images without requiring an extensive dataset or 
exhaustive training.